Safari-609.1.11.1/LayoutTests/accessibility/mac/webarea-can-set-focus-expected.txt

This tests that the web area's focus can be settable.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS body.isFocusable is true
PASS successfullyParsed is true

TEST COMPLETE
